+++ src/main/kotlin/com/tezos/lang/michelson/editor/stack/actions/HighlightUnchangedAction.kt
@@ -10,13 +10,18 @@ import com.tezos.intellij.ui.Icons
  * @author jansorg
  */
 class HighlightUnchangedAction : ToggleAction("Highlight changes", "Highlights unchanged stack elements.", Icons.StackUnchanged), DumbAware, RightAlignedToolbarAction {
+    override fun update(e: AnActionEvent) {
+        super.update(e)
+        e.presentation.isEnabledAndVisible = SplitActionUtil.findStackEditor(e) != null
+    }
+
     override fun setSelected(e: AnActionEvent, state: Boolean) {
-        val editor = SplitActionUtil.findStackEditor(e)
+        val editor = SplitActionUtil.findStackEditor(e) ?: return
         editor.stackHighlightUnchanged = state
         editor.triggerStackUpdate()
     }
 
     override fun isSelected(e: AnActionEvent): Boolean {
-        return SplitActionUtil.findStackEditor(e).stackHighlightUnchanged
+        return SplitActionUtil.findStackEditor(e)?.stackHighlightUnchanged ?: false
     }
 }