+++ src/main/kotlin/com/tezos/lang/michelson/editor/stack/actions/ShowAnnotationsAction.kt
@@ -10,13 +10,18 @@ import com.tezos.intellij.ui.Icons
  * @author jansorg
  */
 class ShowAnnotationsAction : ToggleAction("Show annotations", "Show annotations", Icons.StackAnnotations), DumbAware, RightAlignedToolbarAction {
+    override fun update(e: AnActionEvent) {
+        super.update(e)
+        e.presentation.isEnabledAndVisible = SplitActionUtil.findStackEditor(e) != null
+    }
+
     override fun setSelected(e: AnActionEvent, state: Boolean) {
-        val editor = SplitActionUtil.findStackEditor(e)
+        val editor = SplitActionUtil.findStackEditor(e) ?: return
         editor.stackShowAnnotations = state
         editor.triggerStackUpdate()
     }
 
     override fun isSelected(e: AnActionEvent): Boolean {
-        return SplitActionUtil.findStackEditor(e).stackShowAnnotations
+        return SplitActionUtil.findStackEditor(e)?.stackShowAnnotations ?: false
     }
 }