+++ src/main/kotlin/com/tezos/lang/michelson/editor/completion/provider/SectionCompletion.kt
@@ -5,21 +5,22 @@ import com.intellij.codeInsight.completion.CompletionProvider
 import com.intellij.codeInsight.completion.CompletionResultSet
 import com.intellij.codeInsight.completion.CompletionType
 import com.intellij.codeInsight.lookup.LookupElementBuilder
+import com.intellij.psi.util.PsiTreeUtil
 import com.intellij.util.ProcessingContext
-import com.tezos.lang.michelson.psi.MichelsonPsiFile
+import com.tezos.lang.michelson.psi.PsiContract
 import com.tezos.lang.michelson.psi.PsiSectionType
 
 internal class SectionCompletion : CompletionProvider<CompletionParameters>() {
     override fun addCompletions(parameters: CompletionParameters, context: ProcessingContext, result: CompletionResultSet) {
         val skippedSections: Set<PsiSectionType> = when (parameters.completionType) {
             CompletionType.SMART -> {
-                val sections = (parameters.position.containingFile as? MichelsonPsiFile)?.getContract()?.sections
-                sections?.map { it.sectionType }?.toSet() ?: emptySet()
+                val contract = PsiTreeUtil.findFirstParent(parameters.position) { it is PsiContract } as? PsiContract
+                contract?.sections?.map { it.sectionType }?.toSet() ?: emptySet()
             }
             else -> emptySet()
         }
 
-        PsiSectionType.values().filter { it !in skippedSections }.forEach {
+        PsiSectionType.completionValues.filter { it !in skippedSections }.forEach {
             val name = it.codeName()
             val item = LookupElementBuilder.create("$name ")
                     .withPresentableText(name)