+++ Changelog.md
@@ -1,3 +1,6 @@
+# Version 1.0.0
+- Don't suggest section names inside of sections or types
+
 # Version 0.10.0
 - Feature: Run configurations for Michelson files. You will be asked for the input values by default.
 - Show an error instead of failing silently when no tezos client is configured